+++ source-api/src/main/kotlin/eu/kanade/tachiyomi/source/model/SChapter.kt
@@ -3,6 +3,7 @@
 package eu.kanade.tachiyomi.source.model
 
 import java.io.Serializable
+import kotlinx.serialization.json.JsonObject
 
 interface SChapter : Serializable {
 
@@ -16,12 +17,15 @@ interface SChapter : Serializable {
 
     var scanlator: String?
 
+    var memo: JsonObject
+
     fun copyFrom(other: SChapter) {
         name = other.name
         url = other.url
         date_upload = other.date_upload
         chapter_number = other.chapter_number
         scanlator = other.scanlator
+        memo = other.memo
     }
 
     companion object {